ml/requirements-ml.txt
numpy
Pillow
tqdm
scikit-learn
matplotlib
tensorflow